Move err cog tests to own file
