+++ tests/test_game_commands_cog.py
@@ -1,10 +1,7 @@
-import copy
-import re
 from unittest.mock import AsyncMock
 
 import pytest
 from callee import Regex, String
-from discord.ext import commands
 
 from word_debt_bot import cogs
 from word_debt_bot import game as game_lib
@@ -45,25 +42,6 @@ async def test_registration(
 
 
 @pytest.mark.asyncio
-async def test_registration_valueerror(
-    bot: client.WordDebtBot, player: game_lib.WordDebtPlayer
-):
-    ctx = AsyncMock()
-    ctx.author.id = player.user_id
-    ctx.author.name = player.display_name
-    game_cmds_cog = bot.get_cog("Core Gameplay")
-    cmd_err_cog = bot.get_cog("Command Error Handler")
-
-    await game_cmds_cog.register.invoke(ctx)
-    with pytest.raises(commands.CommandInvokeError) as err:
-        await game_cmds_cog.register.invoke(ctx)
-
-    await cmd_err_cog.on_command_error(ctx, err.value)
-
-    ctx.send.assert_called_with("Already registered!")
-
-
-@pytest.mark.asyncio
 async def test_submit_words(
     game_commands_cog: cogs.GameCommands, player: game_lib.WordDebtPlayer
 ):
@@ -77,94 +55,12 @@ async def test_submit_words(
 
 
 @pytest.mark.asyncio
-async def test_submit_words_with_error(
-    bot: client.WordDebtBot, player: game_lib.WordDebtPlayer
-):
-    ctx = AsyncMock()
-    ctx.author.id = player.user_id
-    game_cmds_cog = bot.get_cog("Core Gameplay")
-    cmd_err_cog = bot.get_cog("Command Error Handler")
-    game_cmds_cog.game.register_player(player)
-
-    @bot.command()
-    async def log_test_m1k(ctx, words=-1000):
-        await game_cmds_cog.log(ctx, words, genre=None)
-
-    game_cmds_cog.log_test_m1k = log_test_m1k
-
-    with pytest.raises(commands.CommandInvokeError) as err:
-        await game_cmds_cog.log_test_m1k.invoke(ctx)
-
-    ctx.command.name = "log"
-
-    await cmd_err_cog.on_command_error(ctx, err.value)
-
-    ctx.send.assert_called_with(String() & Regex("Error:.*"))
-
-
-@pytest.mark.asyncio
-async def test_submit_words_with_no_register(
-    bot: client.WordDebtBot, player: game_lib.WordDebtPlayer
-):
-    ctx = AsyncMock()
-    ctx.author.id = player.user_id
-    game_cmds_cog = bot.get_cog("Core Gameplay")
-    cmd_err_cog = bot.get_cog("Command Error Handler")
-
-    @bot.command()
-    async def log_test_10(ctx, words=10):
-        await game_cmds_cog.log(ctx, words, None)
-
-    game_cmds_cog.log_test_10 = log_test_10
-
-    with pytest.raises(commands.CommandInvokeError) as err:
-        await game_cmds_cog.log_test_10.invoke(ctx)
-
-    ctx.command.name = "log"
-
-    await cmd_err_cog.on_command_error(ctx, err.value)
-
-    ctx.send.assert_called_with(String() & Regex("Not registered!.*"))
-
-
-@pytest.mark.asyncio
 async def test_request_leaderboard_no_register(game_commands_cog: cogs.GameCommands):
     ctx = AsyncMock()
 
     await game_commands_cog.leaderboard(game_commands_cog, ctx, "debt", 1)
 
     ctx.send.assert_called_with(String() & Regex("No registered users.*"))
-
-
-@pytest.mark.asyncio
-async def test_buy_with_value_error(
-    bot: client.WordDebtBot, player: game_lib.WordDebtPlayer
-):
-    ctx = AsyncMock()
-    ctx.author.id = player.user_id
-    target_player = copy.copy(player)
-    target_player.user_id = "12345"
-    game_cmds_cog = bot.get_cog("Core Gameplay")
-    cmd_err_cog = bot.get_cog("Command Error Handler")
-    game_cmds_cog.game.register_player(player)
-    game_cmds_cog.game.register_player(target_player)
-
-    @bot.command()
-    async def buy_test_err(ctx):
-        await game_cmds_cog.buy(
-            ctx, "debt increase", args=f"<@{target_player.user_id}>"
-        )
-
-    game_cmds_cog.buy_test_err = buy_test_err
-
-    with pytest.raises(commands.CommandInvokeError) as err:
-        await game_cmds_cog.buy_test_err.invoke(ctx)
-
-    ctx.command.name = "buy"
-
-    await cmd_err_cog.on_command_error(ctx, err.value)
-
-    ctx.send.assert_called_with("Error: insufficient cranes")
 
 
 @pytest.mark.asyncio
